--- components/Cheese/Cheese.css
@@ -1,50 +1,0 @@
-@import "../palette.css";
-
-.container {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  flex-wrap: wrap;
-  margin-top: 15px;
-
-  @media only screen and (max-width: 650px) {
-    flex-direction: column;
-  }
-}
-
-.section {
-  /* this actually better than calc */
-  width: 33.3333%;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: center;
-
-  @media only screen and (max-width: 650px) {
-    width: 100%;
-    margin-bottom: 25px;
-  }
-}
-
-/* Make button aligned to center of container due to huge padding at the bottom */
-.section:nth-child(3) {
-  justify-content: center;
-}
-
-.cheeseTitle {
-  font-weight: 700;
-  margin-top: 0;
-  margin-bottom: 10px;
-}
-
-/* I saved class on both p tags for possibility to future theming. */
-.cheesePromo:first-child{
-  margin: 0;
-}
-
-.content {
-  width: 100%;
-  height: 100%;
-  text-align: center;
-  /*text-shadow: 0px 1px 2px #000000;*/
-}
